requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-decouple==3.7
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
